fix java Bmi

+++ src/BmiService.java
@@ -1,5 +1,6 @@
 public class BmiService {
-    public double calculate( double BodyMass, double height) {
+    public double calculate(double BodyMass, double height) {
         double BodyMassIndex = BodyMass / (height * height);
-        return (double) BodyMassIndex; }
+        return BodyMassIndex;
+    }
 }
